fix: test 완료 및 test 폴더 삭제

+++ gc-coffee/src/main/java/hello/gccoffee/entity/Product.java
@@ -20,7 +20,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long productId;
+    private Integer productId;
 
     @Column(length = 100)
     private String productName;
@@ -28,7 +28,7 @@ public class Product {
     @Enumerated(EnumType.STRING)
     private Category category;
 
-    private int price;
+    private Integer price;
 
     @Column(length = 2000)
     private String description;
